refactor: move repositories from service/ to data/remote/
Separate data access repositories from Android services:
- OpenAiRepository → data/remote/
- TranslateRepository → data/remote/
- InstapaperRepository → data/remote/
- GeminiData → data/remote/model/

service/ now only contains actual Android services (ClearService,
TtsManager, TtsNotificationManager, TtsNotificationReceiver).
